apply pr feedback
Radically simplify a consitent way to building a plugin for ENSIndexer.

Focuses on reducing meta-programming and replacing it with explicit code structures. For example, all event handlers for a plugin are included by a single function call. No arguments are requried anymore to attach event listeners for a plugin. All required params are provided inline, within relevant event handlers scope.