deps(ponder): change Ponder Status sourcing method
Ponder Status database API has been replaced with Ponder Checkpoints database API which is hard to use on the application side. We use a workaround which makes Ponder instance to make the `GET /status` HTTP request to itself. We parse the response and return the validated result for the application to use.