fix(ensapi): use native promises with swr cache
The SWR cache requires `fn` to return a native Promise object, so we cannot wrap the returned value with `pReflect`. On the other hand, in order to maintain the downstream logic which relies on `pReflect` result type, we still use `pReflect`, but this time on the middleware level, and not on the cache query fn level.