fix: Referrer Leaderboard caching
The fix requires dropping support for `accurateAsOf` field. Value of this field needs to be based on the cached indexing status. This breaks the global cache scope guarantees, and makes the SWR cache for Referrer Leaderboard scoped to the current request. This is unacceptable behaviour so we decided to drop `accurateAsOf` field from data model to remove Referrer Leaderboard cache dependency on the cached indexing status.