chore: Updates to name-tokens-api types
The Hono OpenAPI implementation with Zod produces a JSON schema where
`z.undefined()` is
[unrepresentable](https://zod.dev/json-schema#unrepresentable). This
commit refactors the `NameTokensRequest` type and by consequence update
the zod schema to pass. In order for OpenAPI JSON schema to work we
cannot have any instances of unrepresentable APIs, and this so far was
the only case.